requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pdfkit==1.0.0
qrcode[pil]==7.4.2